+++ petshopMentoria/app/Http/Controllers/api/AgendamentoController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\Agendamento\AgendamentoPostRequest;
 use App\Repositories\Contracts\AgendamentoRepositoryInterface;
 use App\Repositories\Contracts\FuncionarioRepositoryInterface;
+use App\Services\AgendamentoService;
 use Facade\FlareClient\Http\Response;
 use Fig\Http\Message\StatusCodeInterface;
 use Illuminate\Http\Request;
@@ -14,100 +15,35 @@ use function PHPSTORM_META\type;
 
 class AgendamentoController extends Controller
 {
-    private $repository;
-    private $funcionario;
+    private $service;
 
-    public function __construct(AgendamentoRepositoryInterface $repository, FuncionarioRepositoryInterface $funcionarioRepository)
+    public function __construct(AgendamentoService $service)
     {
-        $this->repository = $repository;
-        $this->funcionario = $funcionarioRepository;
+        $this->service = $service;
     }
 
     public function index()
     {
-        $agenda = $this->repository->all();
-        return Response($agenda, StatusCodeInterface::STATUS_OK);
+        return $this->service->index();
     }
 
     public function post(AgendamentoPostRequest $request)
     {
-        $funcionario = $this->funcionario->find($request->get('funcionario_id'));
-        $servico = $request->get('servico_id');
-
-        if ($funcionario->fazServico($servico)) {
-            if ($funcionario->disponivel($request->get('inicio'), $request->get('fim'))) {
-                $agendamento = $this->repository->create($request->all());
-                return Response($agendamento, StatusCodeInterface::STATUS_CREATED);
-            } else {
-                return Response(
-                    ['erro' => 'Este colaraborador não está disponível para este horário.'],
-                    StatusCodeInterface::STATUS_NOT_FOUND
-                );
-            }
-        } else {
-            return Response(
-                ['erro' => 'Este colaraborador não realiza este tipo de serviço.'],
-                StatusCodeInterface::STATUS_NOT_FOUND
-            );
-        }
+        return $this->service->post($request);
     }
 
     public function show(int $id)
     {
-        $agenda = $this->repository->find($id);
-
-        if ($agenda) {
-            return Response($agenda, StatusCodeInterface::STATUS_OK);
-        }
-        return Response([], StatusCodeInterface::STATUS_NOT_FOUND);
+        return $this->service->show($id);
     }
 
     public function put(AgendamentoPostRequest $request, int $id)
     {
-        $funcionario = $this->funcionario->find($request->get('funcionario_id'));
-        $servico = $request->get('servico_id');
-
-        if ($funcionario->fazServico($servico)) {
-            if (
-                $funcionario->disponivel($request->get('inicio'), $request->get('fim'), $id)) {
-                if ($this->repository->update($request->all(), $id)) {
-                    return Response(
-                        ['status' => 'Recurso atualizado com sucesso.'],
-                        StatusCodeInterface::STATUS_OK
-                    );
-                }
-
-                return Response(
-                    [],
-                    StatusCodeInterface::STATUS_NOT_FOUND
-                );
-            } else {
-                return Response(
-                    ['erro' => 'Este colaraborador não está disponível para este horário.'],
-                    StatusCodeInterface::STATUS_NOT_FOUND
-                );
-            }
-        } else {
-            return Response(
-                ['erro' => 'Este colaraborador não realiza este tipo de serviço.'],
-                StatusCodeInterface::STATUS_NOT_FOUND
-            );
-        }
+        return $this->service->put($request, $id);
     }
 
     public function delete(int $id)
     {
-        return Response($this->repository->find($id));
-        if ($this->repository->delete($id)) {
-            return Response(
-                [],
-                StatusCodeInterface::STATUS_NO_CONTENT
-            );
-        }
-
-        return Response(
-            [],
-            StatusCodeInterface::STATUS_NOT_FOUND
-        );
+        return $this->service->delete($id);
     }
 }
